+++ venue_review_api/src/controllers/venue.controller.ts
@@ -110,12 +110,12 @@ const getVenues = async (req: Request, res: Response) => {
 
   let maxCostRating =
     req.query.maxCostRating != null &&
-    req.query.maxCostRating.toString().length > 0
+      req.query.maxCostRating.toString().length > 0
       ? Number(req.query.maxCostRating?.toString())
       : null;
   let minStarRating =
     req.query.minStarRating != null &&
-    req.query.minStarRating.toString().length > 0
+      req.query.minStarRating.toString().length > 0
       ? Number(req.query.minStarRating?.toString())
       : null;
 
@@ -263,7 +263,7 @@ const createVenuePhoto = async (req: Request, res: Response) => {
   let fileName = `${venue_id}-${uuidv4().replace(/-/g, '')}.${imageExt}`;
 
   let imageDIR = `./${venuePhotoBucket}`;
-  if (!fs.existsSync(imageDIR)) {
+  if (!fs.existsSync(imageDIR) && process.env.NODE_ENV != 'test') {
     fs.mkdirSync(imageDIR);
   }
 
@@ -298,7 +298,9 @@ const createVenuePhoto = async (req: Request, res: Response) => {
         });
     });
   } catch (err) {
-    fs.rmSync(imageDIR, { recursive: true }); // Delete the local file, we had a failure, and don't want these to hang around.
+    // Only do real file deletes if we are not in test mode.
+    if (process.env.NODE_ENV != 'test')
+      fs.rmSync(imageDIR, { recursive: true }); // Delete the local file, we had a failure, and don't want these to hang around.
     res.status(500).json({ status: 500, message: err });
   }
 };
@@ -360,5 +362,6 @@ export {
   getVenues,
   removePhoto,
   setNewPrimary,
-  updateVenue,
+  updateVenue
 };
+
